requirements.txt
Adafruit-BBIO (0.0.20)
pip (1.5.6)
pyserial (2.7)
setuptools (3.6)
wsgiref (0.1.2)
